+++ core-services/pgr-services/src/main/java/org/egov/pgr/web/models/Boundary.java
@@ -6,11 +6,10 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.validator.constraints.SafeHtml;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,7 +17,7 @@ import java.util.List;
  * Boundary
  */
 @Validated
-@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
+@jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
 
 @Getter
 @Setter
@@ -28,23 +27,18 @@ import java.util.List;
 public class Boundary   {
 
         @NotNull
-        @SafeHtml
         @JsonProperty("code")
         private String code = null;
 
-        @SafeHtml
         @JsonProperty("name")
         private String name = null;
 
-        @SafeHtml
         @JsonProperty("label")
         private String label = null;
 
-        @SafeHtml
         @JsonProperty("latitude")
         private String latitude = null;
 
-        @SafeHtml
         @JsonProperty("longitude")
         private String longitude = null;
 
@@ -52,7 +46,6 @@ public class Boundary   {
         @Valid
         private List<Boundary> children = null;
 
-        @SafeHtml
         @JsonProperty("materializedPath")
         private String materializedPath = null;