+++ core-services/pgr-services/src/main/java/org/egov/pgr/web/models/Document.java
@@ -7,7 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.validator.constraints.SafeHtml;
 import org.springframework.validation.annotation.Validated;
 
 /**
@@ -15,7 +14,7 @@ import org.springframework.validation.annotation.Validated;
  */
 @ApiModel(description = "This object holds list of documents attached during the transaciton for a property")
 @Validated
-@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
+@jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
 
 @Getter
 @Setter
@@ -23,19 +22,15 @@ import org.springframework.validation.annotation.Validated;
 @NoArgsConstructor
 @Builder
 public class Document   {
-        @SafeHtml
         @JsonProperty("id")
         private String id = null;
 
-        @SafeHtml
         @JsonProperty("documentType")
         private String documentType = null;
 
-        @SafeHtml
         @JsonProperty("fileStoreId")
         private String fileStoreId = null;
 
-        @SafeHtml
         @JsonProperty("documentUid")
         private String documentUid = null;